24/10/2023

+++ app/Http/Controllers/PhieunhapxuatController.php
@@ -12,7 +12,6 @@ use App\Models\CTnhapxuat;
 use App\Models\Nhansu;
 use App\Models\DCnhapxuat;
 use App\Models\Trangthai;
-use App\Models\Kho;
 use App\Models\Tonkho;
 
 
@@ -21,10 +20,12 @@ class PhieunhapxuatController extends Controller
     public function index(){   // Giao diện hiển thị toàn bộ dữ liệu: GET
 
         $phieu = Phieunhapxuat::all();
+        $Trangthai = Trangthai::all();
 
         return view("giaodien.app", [
             'page' => "phieunhapxuat.DSphieu",
-            'phieu' => $phieu
+            'phieu' => $phieu,
+            'Trangthai'=> $Trangthai
         ]);
     }
 
@@ -190,7 +191,7 @@ class PhieunhapxuatController extends Controller
 
             }
 
-            return back()->with('alert', 'Thêm dữ liệu thành công');
+            return back()->with('alert', 'Tạo phiếu '.$Trangthai->TENTT.' thành công');
 
         } catch (Exception $err) {
             return back()->withError($err->getMessage())->withInput();
@@ -199,32 +200,35 @@ class PhieunhapxuatController extends Controller
     }
 
     public function show($id){   // Lấy chi tiết của một dữ liệu: GET
+
         $phieu = Phieunhapxuat::find($id);
+        $ctphieu = CTnhapxuat::where('MAPHIEU', $id);
 
         return view("giaodien.app", [
             'page' => "phieunhapxuat.DSphieu",
-            'phieu' => $phieu
+            'phieu' => $phieu,
+            'ctphieu'=> $ctphieu
         ]);
     }
 
     public function edit($id){   // Giao diện cập nhật dữ liệu: GET
 
         $phieu = Phieunhapxuat::find($id);
+        $DCnhapxuat = DCnhapxuat::all();
 
         return view("giaodien.app", [
             'page' => "phieunhapxuat.ViewUpdataPhieu",
-            'phieu' => $phieu
+            'phieu' => $phieu,
+            'DCnhapxuat' => $DCnhapxuat
         ]);
     }
 
-    public function update(Request $request, $id){   // Cập nhật lại dữ liệu: POST
+    public function update(Request $request){   // Cập nhật lại dữ liệu: POST
         
         $rules = [
             'sophieu' => 'required|alpha|max:20',
             'madiachi' => 'required|numeric',
-            'matrangthai' => 'required|numeric',
-            'sp.*.slsp' => 'required|numeric',
-            'sp.*.idsp' => 'required|numeric|distinct:strict'
+            'idphieu'=> 'required|numeric'
         ];
 
         $mess = [
@@ -232,131 +236,19 @@ class PhieunhapxuatController extends Controller
             'sophieu.alpha' => 'Vui lòng nhập ký tự chữ cái',
             'sophieu.max' => 'Tối đa 20 kí tự',
             'madiachi.required' => 'Chưa chọn thông tin',
-            'matrangthai.required' => 'Chưa chọn thông tin',
-            'sp.*.idsp.required' => 'Chưa chọn thông tin',
-            'sp.*.idsp.distinct' => 'Không được trùng lặp thông tin',
-            'sp.*.slsp.required' => 'Vui lòng nhập số lượng',
-            'sp.*.slsp.numeric' => 'Vui lòng nhập chữ số',
+            'idphieu.required'=> 'Chưa có id thông tin'
         ];
 
         $request->validate($rules, $mess);
 
         try {
 
-            $sophieu = $request->sophieu;
-
-            $idnhanvien = Auth::id();
-
-            $nhanvien =  Nhansu::firstWhere('MANV', $idnhanvien);
-
-            $Trangthai = Trangthai::find($request->matrangthai);
-
-            $makho = $nhanvien->MAKHO;
-
-            // cập nhật tồn kho
-            switch ($Trangthai->TENTT) {
-
-                case 'Xuất':
-                    
-                    foreach ($request->sp as $key => $value) {
-
-                        $tonkho = Tonkho::where('MAKHO', $makho)
-                        ->where('MASP', $value['idsp'])
-                        ->first();
-
-                        if (empty($tonkho)) {
-
-                            $sp = Sanpham::find($value['idsp']);
-
-                            Tonkho::create([
-                                'MAKHO' => $makho,
-                                'MASP' => $value['idsp'],
-                                'SLTONKHO' => 0,
-                                'SLNHAP' => 0,
-                                'SLXUAT' => 0
-                            ]);
-
-                            return back()->with('err', 'Hiện chưa có sản phẩm '.$sp->TENSP.' trong kho, vui lòng nhập hàng');
-                        }
-                        else {
-
-                            if ($tonkho->SLTONKHO < $value['slsp'] || $value['slsp'] <= 0) {
-
-                                $sp = Sanpham::find($value['idsp']);
-    
-                                return back()->with('err', 'Hiện chỉ còn '.$tonkho->SLTONKHO.' sản phẩm '.$sp->TENSP);
-                            }
-    
-                            Tonkho::where('MAKHO', $makho)
-                            ->where('MASP', $value['idsp'])
-                            ->update([
-                                'SLTONKHO' => $tonkho->SLTONKHO - $value['slsp'],
-                                'SLXUAT' => $tonkho->SLXUAT + $value['slsp'],
-                            ]);
-                        }
-
-                    }
-
-                    break;
-                
-                default:
-                    
-                    foreach ($request->sp as $key => $value) {
-
-                        $tonkho = Tonkho::where('MAKHO', $makho)
-                        ->where('MASP', $value['idsp'])
-                        ->first();
-
-                        if (empty($tonkho)) {
-                            Tonkho::create([
-                                'MAKHO' => $makho,
-                                'MASP' => $value['idsp'],
-                                'SLTONKHO' => $tonkho->SLTONKHO + $value['slsp'],
-                                'SLNHAP' => $tonkho->SLNHAP + $value['slsp'],
-                                'SLXUAT' => 0
-                            ]);
-                        }
-                        else {
-                            Tonkho::where('MAKHO', $makho)
-                            ->where('MASP', $value['idsp'])
-                            ->update([
-                                'SLTONKHO' => $tonkho->SLTONKHO + $value['slsp'],
-                                'SLNHAP' => $tonkho->SLNHAP + $value['slsp'],
-                            ]);
-                        }
-
-                    }
-
-                    break;
-            }
-
             // cập nhật phiếu và chi tiết phiếu
-            Phieunhapxuat::where('id', $id)
+            Phieunhapxuat::where('id', $request->idphieu)
                             ->update([
-                                'SOPHIEU' => $request->sophieu,
-                                'MADC' => $request->madiachi,
-                                'MATT' => $request->matrangthai
+                                'SOPHIEU' => strtoupper($request->sophieu).'-'.$request->idphieu,
+                                'MADC' => $request->madiachi
                             ]);
-    
-            Phieunhapxuat::where('id', $id)->update(['SOPHIEU' => strtoupper($sophieu).'-'.$id]);
-
-            foreach ($request->sp as $key => $value) {
-
-                $sl = $value['slsp'];
-                $sp = Sanpham::find($value['idsp']);
-                $gia = $sp->GIASP;
-                $dongia = $gia * $sl;
-
-                CTnhapxuat::where('MAPHIEU', $id)
-                            ->where('MASP', $value['idsp'])
-                            ->update([
-                                'SOLUONG' => $sl, 
-                                'DONGIA' => $gia, 
-                                'THANHTIEN' => $dongia, 
-                                'MAPHIEU' => $id, 
-                            ]);
-
-            }
 
             return back()->with('alert', 'Cập nhật thành công');
 
@@ -366,8 +258,25 @@ class PhieunhapxuatController extends Controller
 
     }
 
-    public function destroy($id){   // Xóa bỏ một dữ liệu: GET
-        $phieu = Phieunhapxuat::find($id);
-        $phieu->delete();
+    public function destroy(Request $request){   // Xóa bỏ một dữ liệu: GET
+
+        $rules = [
+            'idphieu' => 'required|numeric'
+        ];
+
+        $mess = [
+            'idphieu.required' => 'Chưa có id thông tin'
+        ];
+
+        $request->validate($rules, $mess);
+
+        $phieu = Phieunhapxuat::find($request->idphieu);
+
+        if($phieu){
+            $phieu->delete();
+            return back()->with('alert', 'Xóa thành công');
+        }
+
+        return back()->with('err', 'Không tồn tại dữ liệu cần xóa');
     }
 }
